pyproject.toml
[tool.poetry]
name = "licence-plates-recognition"
version = "0.1.0"
description = ""
repository = "https://github.com/sn09/licence_plates_recognition"
license = "MIT Licence"
authors = ["Saraev Nikita <saraevnik0909@gmail.com>"]
readme = "README.md"
packages = [
    {include = "recognition_app"}
]


[tool.poetry.dependencies]
python = "^3.10"
opencv-python = "^4.8.1.78"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
torch = "^2.1.2"
torchvision = "^0.16.2"
pandas = "^2.1.4"
scikit-learn = "^1.3.2"
dvc = {extras = ["s3"], version = "^3.38.1"}
omegaconf = "^2.3.0"
lightning = "^2.1.3"
mlflow = "^2.9.2"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pre-commit = "^3.6.0"
ipykernel = "^6.27.1"


[tool.black]
line-length = 90
target-version = ["py310"]


[tool.isort]
src_paths = ["recognition_app"]
profile = "black"
line_length = 90
lines_after_imports = 2


[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = ".flake8"


[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]


[tool.nbqa.mutate]
black = 1
isort = 1


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
